+++ messaging/src/resultado_linea.rs
@@ -5,7 +5,7 @@ pub enum ResultadoLinea {
     Pub(String, Option<String>, usize),
     Hpub(String, Option<String>, usize, usize),
     Sub(String, Option<String>, String),
-    Unsub(String, Option<usize>),
+    Unsub(String, Option<u64>),
     Ping,
     Connect
 }
